refactored option type patching to new service
