+++ src/app/modules/local-admin/local-admin-organization/local-admin-organization.component.html
@@ -6,7 +6,7 @@
 
   <app-option-type-table
     *ngIf="selectedSegment === LocalAdminOrganizationSegmentOption.Roles"
-    optionType="organization-unit"
+    optionType="it-system_business-type"
     title="Roller i din orginsation"
     i18n-title
     [disableAccordion]="true"
