+++ src/app/shared/components/option-type-table/edit-option-type-dialog/edit-option-type-dialog.component.scss
@@ -1,5 +1,2 @@
 @import "theme";
 
-app-standard-vertical-content-grid * {
-  padding: $padding-small;
-}
