+++ src/styles/theme.scss
@@ -52,6 +52,7 @@ $light-primary-text: $color-white;
 
 // Sizes
 
+$padding-smallest: 2px;
 $padding-smaller: 4px;
 $padding-small: 8px;
 $padding-medium: 16px;
